Expose CloseAndReconnect method
The method seems to have been accidentally not exposed since
the comment is uppercase but the function name is lowercase
and all the other methods seems to match their comment and
function name casing.

Even if not, this can be useful to the users - a server might send
some kind of a `reconnect` order that dictates the client to
terminate the connection and reconnect.