dial: return early on successful initial dial instead of blocking via sleep unconditionally
Dial should not unconditionally sleep for that amount of time. If the conection succeeded via the asynchronous goroutine, we can simply return early.

This is particularly helpful in situations where one needs a higher handshake timeout, because now Dial won't just block for the full amount of time, allowing the program to move on to other work.